pyproject.toml
[tool.poetry]
name = "upcy-be"
version = "0.1.0"
description = ""
authors = ["Scanf-s <calzone0404@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1.1"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.2.2"
django-extensions = "^3.2.3"
pillow = "^10.4.0"
gunicorn = "^23.0.0"
uvicorn = "^0.30.6"
boto3 = "^1.35.19"
django-cors-headers = "^4.4.0"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
flake8 = "^7.1.1"
black = "^24.8.0"
mypy = "^1.11.2"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
